Move role controller to standard Teleport controller
